requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
pix2text==1.1.4
pillow==12.0.0